Changed column types from integer to double.

+++ app/database/migrations/2013_10_31_190858_create_willingness_to_pay_treatment_table.php
@@ -18,7 +18,7 @@ class CreateWillingnessToPayTreatmentTable extends Migration {
 		{
             $table->increments(WillingnessPay::$ID_KEY);
 			$table->integer(WillingnessPay::$SESSION_ID_KEY)->unsigned();
-            $table->integer(WillingnessPay::$ENDOWMENT_KEY);
+            $table->double(WillingnessPay::$ENDOWMENT_KEY);
 		});
 	}
